add diff structure for siafunds
